# swagger.yml
openapi: 3.0.3
info:
  title: Online Subscription Service API
  description: API для управления подписками пользователей и подсчета расходов.
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local server

paths:
  /api/v1/subscription:
    post:
      summary: Создание подписки
      tags:
        - subscriptions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubscriptionRequest'
      responses:
        '201':
          description: Подписка создана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionRequest'
        '400':
          description: Ошибка валидации данных
        '500':
          description: Внутренняя ошибка сервера

    get:
      summary: Получить список всех подписок
      tags:
        - subscriptions
      responses:
        '200':
          description: Список подписок
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SubscriptionRequest'
        '500':
          description: Внутренняя ошибка сервера

  /api/v1/subscription/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
        description: UUID подписки
    get:
      summary: Получить подписку по ID
      tags:
        - subscriptions
      responses:
        '200':
          description: Данные подписки
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionRequest'
        '400':
          description: Неверный формат ID
        '404':
          description: Подписка не найдена
        '500':
          description: Внутренняя ошибка сервера

    put:
      summary: Обновить подписку
      tags:
        - subscriptions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubscriptionRequest'
      responses:
        '200':
          description: Подписка обновлена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionRequest'
        '400':
          description: Ошибка валидации или ID
        '404':
          description: Подписка не найдена
        '500':
          description: Внутренняя ошибка сервера

    delete:
      summary: Удалить подписку
      tags:
        - subscriptions
      responses:
        '200':
          description: Подписка удалена
          content:
            application/json:
              schema:
                type: object
        '400':
          description: Неверный формат ID
        '404':
          description: Подписка не найдена
        '500':
          description: Внутренняя ошибка сервера

  /api/v1/subscription/total:
    get:
      summary: Получить общую стоимость
      tags:
        - stats
      parameters:
        - name: user_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: start_date
          in: query
          required: true
          description: Format MM-YYYY
          schema:
            type: string
            example: "01-2025"
        - name: end_date
          in: query
          required: true
          description: Format MM-YYYY
          schema:
            type: string
            example: "12-2025"
        - name: service_name
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Итоговая сумма
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatsResponse'
        '400':
          description: Ошибка параметров запроса
        '500':
          description: Ошибка подсчета

components:
  schemas:
    SubscriptionRequest:
      type: object
      required:
        - service_name
        - price
        - user_id
        - start_date
      properties:
        service_name:
          type: string
          example: "Netflix"
        price:
          type: integer
          example: 1000
        user_id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        start_date:
          type: string
          example: "01-2025"
          description: "Format: MM-YYYY"
        end_date:
          type: string
          example: "02-2025"
          description: "Format: MM-YYYY"

    StatsResponse:
      type: object
      properties:
        total_price:
          type: integer
          example: 5000